+++ yash-syntax/src/parser/lex/dollar.rs
@@ -41,12 +41,12 @@ impl WordLexer<'_, '_> {
             Err(location) => location,
         };
 
-        let location = match self.braced_param(location).await? {
+        let _location = match self.braced_param(location).await? {
             Ok(result) => return Ok(Some(TextUnit::BracedParam(result))),
             Err(location) => location,
         };
 
-        if let Some(result) = self.arithmetic_expansion(location).await? {
+        if let Some(result) = self.arithmetic_expansion(start_index).await? {
             return Ok(Some(result));
         }